Squashed commit of v4.6.5

+++ buildscripts/ci/linux/tools/make_appimage.sh
@@ -1,5 +1,7 @@
 #!/usr/bin/env bash
 
+trap 'code=$?; echo "error: make_appimage.sh: command \`$BASH_COMMAND\` exited with code $code." >&2; exit 1' ERR
+
 INSTALL_DIR="$1" # MuseScore was installed here
 APPIMAGE_NAME="$2" # name for AppImage file (created outside $INSTALL_DIR)
 PACKARCH="$3" # architecture (x86_64, aarch64, armv7l)
@@ -22,6 +24,11 @@ mkdir -p $BUILD_TOOLS
 # INSTALL APPIMAGETOOL AND LINUXDEPLOY
 ##########################################################################
 
+if [ "$PACKARCH" == "armhf" ]; then
+  # In a Docker container, AppImages cannot run normally because of problems with FUSE.
+  export APPIMAGE_EXTRACT_AND_RUN=1
+fi
+
 function download_github_release()
 {
   local -r repo_slug="$1" release_tag="$2" file="$3"
@@ -30,20 +37,13 @@ function download_github_release()
   else
     local -r url="https://github.com/${repo_slug}/releases/download/${release_tag}/${file}"
   fi
+
+  echo "try download: ${url}"
+  
   # use curl instead of wget which fails on armhf
   curl "${url}" -O -L
   chmod +x "${file}"
-}
-
-function extract_appimage()
-{
-  # Extract AppImage so we can run it without having to install FUSE
-  local -r appimage="$1" binary_name="$2"
-  local -r appdir="${appimage%.AppImage}.AppDir"
-  "./${appimage}" --appimage-extract >/dev/null # dest folder "squashfs-root"
-  mv squashfs-root "${appdir}" # rename folder to avoid collisions
-  ln -s "${appdir}/AppRun" "${binary_name}" # symlink for convenience
-  rm -f "${appimage}"
+  echo "downloaded: ${file}"
 }
 
 function download_appimage_release()
@@ -51,29 +51,17 @@ function download_appimage_release()
   local -r github_repo_slug="$1" binary_name="$2" tag="$3"
   local -r appimage="${binary_name}-${PACKARCH}.AppImage"
   download_github_release "${github_repo_slug}" "${tag}" "${appimage}"
-  extract_appimage "${appimage}" "${binary_name}"
-  # mv "${appimage}" "${binary_name}" # use this instead of the previous line for the static runtime AppImage
+  mv "${appimage}" "${binary_name}"
 }
 
 if [[ ! -d $BUILD_TOOLS/appimagetool ]]; then
   mkdir $BUILD_TOOLS/appimagetool
   cd $BUILD_TOOLS/appimagetool
-  download_appimage_release AppImage/AppImageKit appimagetool continuous # use AppImage/appimagetool for the static runtime AppImage
+  download_appimage_release AppImage/appimagetool appimagetool continuous
   cd $ORIGIN_DIR
 fi
 export PATH="$BUILD_TOOLS/appimagetool:$PATH"
 appimagetool --version
-
-if [[ ! -d $BUILD_TOOLS/appimageupdatetool ]]; then
-  mkdir $BUILD_TOOLS/appimageupdatetool
-  cd $BUILD_TOOLS/appimageupdatetool
-  download_appimage_release AppImage/AppImageUpdate appimageupdatetool continuous
-  cd $ORIGIN_DIR
-fi
-if [[ "${UPDATE_INFORMATION}" ]]; then
-  export PATH="$BUILD_TOOLS/appimageupdatetool:$PATH"
-  appimageupdatetool --version
-fi
 
 function download_linuxdeploy_component()
 {
@@ -108,16 +96,16 @@ mv "${appdir}/bin/findlib" "${appdir}/../findlib"
 # Remove Qt plugins for MySQL and PostgreSQL to prevent
 # linuxdeploy-plugin-qt from failing due to missing dependencies.
 # SQLite plugin alone should be enough for our AppImage.
-# rm -f ${QT_PATH}/plugins/sqldrivers/libqsql{mysql,psql}.so
-qt_sql_drivers_path="${QT_PATH}/plugins/sqldrivers"
+# rm -f ${QT_ROOT_DIR}/plugins/sqldrivers/libqsql{mysql,psql}.so
+qt_sql_drivers_path="${QT_ROOT_DIR}/plugins/sqldrivers"
 qt_sql_drivers_tmp="/tmp/qtsqldrivers"
 mkdir -p "$qt_sql_drivers_tmp"
-mv "${qt_sql_drivers_path}/libqsqlmysql.so" "${qt_sql_drivers_tmp}/libqsqlmysql.so"
-mv "${qt_sql_drivers_path}/libqsqlpsql.so" "${qt_sql_drivers_tmp}/libqsqlpsql.so"
+[ -f "${qt_sql_drivers_path}/libqsqlmysql.so" ] && mv "${qt_sql_drivers_path}/libqsqlmysql.so" "${qt_sql_drivers_tmp}/libqsqlmysql.so"
+[ -f "${qt_sql_drivers_path}/libqsqlpsql.so" ] && mv "${qt_sql_drivers_path}/libqsqlpsql.so" "${qt_sql_drivers_tmp}/libqsqlpsql.so"
 
 # Semicolon-separated list of platforms to deploy in addition to `libqxcb.so`.
 # Used by linuxdeploy-plugin-qt.
-export EXTRA_PLATFORM_PLUGINS="libqoffscreen.so"
+export EXTRA_PLATFORM_PLUGINS="libqoffscreen.so;libqwayland-egl.so;libqwayland-generic.so"
 
 # Colon-separated list of root directories containing QML files.
 # Needed for linuxdeploy-plugin-qt to scan for QML imports.
@@ -127,21 +115,6 @@ export QML_SOURCES_PATHS=./
 linuxdeploy --appdir "${appdir}" # adds all shared library dependencies
 linuxdeploy-plugin-qt --appdir "${appdir}" # adds all Qt dependencies
 
-# Approximately on June 1, the QtQuick/Controls.2 stopped being deploying 
-# (at that time the linux deploy was updated). 
-# This is a hack, for the deployment of QtQuick/Controls.2 
-if [ ! -f ${appdir}/usr/lib/libQt5QuickControls2.so.5 ]; then
-    cp -r ${QT_PATH}/qml/QtQuick/Controls.2 ${appdir}/usr/qml/QtQuick/Controls.2
-    cp -r ${QT_PATH}/qml/QtQuick/Templates.2 ${appdir}/usr/qml/QtQuick/Templates.2
-    cp ${QT_PATH}/lib/libQt5QuickControls2.so.5 ${appdir}/usr/lib/libQt5QuickControls2.so.5 
-    cp ${QT_PATH}/lib/libQt5QuickTemplates2.so.5 ${appdir}/usr/lib/libQt5QuickTemplates2.so.5 
-fi
-
-# At an unknown point in time, the libqgtk3 plugin stopped being deployed
-if [ ! -f ${appdir}/plugins/platformthemes/libqgtk3.so ]; then
-  cp ${QT_PATH}/plugins/platformthemes/libqgtk3.so ${appdir}/plugins/platformthemes/libqgtk3.so 
-fi
-
 # The system must be used
 if [ -f ${appdir}/lib/libglib-2.0.so.0 ]; then
   rm -f ${appdir}/lib/libglib-2.0.so.0 
@@ -149,9 +122,9 @@ fi
 
 unset QML_SOURCES_PATHS EXTRA_PLATFORM_PLUGINS
 
-# In case this container is reused multiple times, return the moved libraries back
-mv "${qt_sql_drivers_tmp}/libqsqlmysql.so" "${qt_sql_drivers_path}/libqsqlmysql.so"
-mv "${qt_sql_drivers_tmp}/libqsqlpsql.so" "${qt_sql_drivers_path}/libqsqlpsql.so"
+# Return the moved libraries back
+[ -f "${qt_sql_drivers_tmp}/libqsqlmysql.so" ] && mv "${qt_sql_drivers_tmp}/libqsqlmysql.so" "${qt_sql_drivers_path}/libqsqlmysql.so"
+[ -f "${qt_sql_drivers_tmp}/libqsqlpsql.so" ] && mv "${qt_sql_drivers_tmp}/libqsqlpsql.so" "${qt_sql_drivers_path}/libqsqlpsql.so"
 
 # Put the non-RUNPATH binaries back
 mv "${appdir}/../findlib" "${appdir}/bin/findlib"
@@ -166,28 +139,19 @@ function find_library()
   "${appdir}/bin/findlib" "$@"
 }
 
-function fallback_library()
-{
-  # Copy a library into a special fallback directory inside the AppDir.
-  # Fallback libraries are not loaded at runtime by default, but they can
-  # be loaded if it is found that the application would crash otherwise.
-  local library="$1"
-  local full_path="$(find_library "$1")"
-  local new_path="${appdir}/fallback/${library}"
-  mkdir -p "${new_path}" # directory has the same name as the library
-  cp -L "${full_path}" "${new_path}/${library}"
-  # Use the AppRun script to check at runtime whether the user has a copy of
-  # this library. If not then add our copy's directory to $LD_LIBRARY_PATH.
-}
-
 # UNWANTED FILES
 # linuxdeploy or linuxdeploy-plugin-qt may have added some files or folders
 # that we don't want. List them here using paths relative to AppDir root.
 # Report new additions at https://github.com/linuxdeploy/linuxdeploy/issues
 # or https://github.com/linuxdeploy/linuxdeploy-plugin-qt/issues for Qt libs.
 unwanted_files=(
-  # none
+  # https://github.com/musescore/MuseScore/issues/24068#issuecomment-2297823192
+  lib/libwayland-client.so.0
 )
+
+for file in "${unwanted_files[@]}"; do
+  rm -rf "${appdir}/${file}"
+done
 
 # ADDITIONAL QT COMPONENTS
 # linuxdeploy-plugin-qt may have missed some Qt files or folders that we need.
@@ -195,15 +159,45 @@ unwanted_files=(
 # additions at https://github.com/linuxdeploy/linuxdeploy-plugin-qt/issues
 additional_qt_components=(
   plugins/printsupport/libcupsprintersupport.so
+
+  # At an unknown point in time, the libqgtk3 plugin stopped being deployed
+  plugins/platformthemes/libqgtk3.so
+
+  # Wayland support (run with MU_QT_QPA_PLATFORM=wayland to use)
+  plugins/wayland-decoration-client
+  plugins/wayland-graphics-integration-client
+  plugins/wayland-shell-integration
 )
+
+for file in "${additional_qt_components[@]}"; do
+  if [ -f "${appdir}/${file}" ]; then
+    echo "Warning: ${file} was already deployed. Skipping."
+    continue
+  fi
+  mkdir -p "${appdir}/$(dirname "${file}")"
+  cp -Lr "${QT_ROOT_DIR}/${file}" "${appdir}/${file}"
+done
 
 # ADDITIONAL LIBRARIES
 # linuxdeploy may have missed some libraries that we need
 # Report new additions at https://github.com/linuxdeploy/linuxdeploy/issues
-additional_libraries=(
-  libssl.so.1.1       # OpenSSL (for Save Online)
-  libcrypto.so.1.1    # OpenSSL (for Save Online)
-)
+if [[ "$PACKARCH" == "x86_64" ]]; then
+  additional_libraries=(
+    libssl.so.1.1    # OpenSSL (for Save Online)
+    libcrypto.so.1.1 # OpenSSL (for Save Online)
+  )
+else
+  additional_libraries=()
+fi
+
+for lib in "${additional_libraries[@]}"; do
+  if [ -f "${appdir}/lib/${lib}" ]; then
+    echo "Warning: ${lib} was already deployed. Skipping."
+    continue
+  fi
+  full_path="$(find_library "${lib}")"
+  cp -L "${full_path}" "${appdir}/lib/${lib}"
+done
 
 # FALLBACK LIBRARIES
 # These get bundled in the AppImage, but are only loaded if the user does not
@@ -215,84 +209,104 @@ additional_libraries=(
 # Report new additions at https://github.com/linuxdeploy/linuxdeploy/issues
 fallback_libraries=(
   libjack.so.0 # https://github.com/LMMS/lmms/pull/3958
+  libOpenGL.so.0 # https://bugreports.qt.io/browse/QTBUG-89754
+  libpipewire-0.3.so.0
 )
-
-# PREVIOUSLY EXTRACTED APPIMAGES
-# These include their own dependencies. We bundle them uncompressed to avoid
-# creating a double layer of compression (AppImage inside AppImage).
-if [[ "${UPDATE_INFORMATION}" ]]; then
-extracted_appimages=(
-  appimageupdatetool
-)
-else
-extracted_appimages=(
-  # none
-)
-fi
-
-for file in "${unwanted_files[@]}"; do
-  rm -rf "${appdir}/${file}"
-done
-
-for file in "${additional_qt_components[@]}"; do
-  mkdir -p "${appdir}/$(dirname "${file}")"
-  cp -Lr "${QT_PATH}/${file}" "${appdir}/${file}"
-done
-
-for lib in "${additional_libraries[@]}"; do
-  full_path="$(find_library "${lib}")"
-  cp -L "${full_path}" "${appdir}/lib/${lib}"
-done
 
 for fb_lib in "${fallback_libraries[@]}"; do
-  fallback_library "${fb_lib}"
+  full_path="$(find_library "$fb_lib")"
+  new_path="${appdir}/fallback/${fb_lib}"
+  mkdir -p "${new_path}" # directory has the same name as the library
+  cp -L "${full_path}" "${new_path}/${fb_lib}"
+  # Use the AppRun script to check at runtime whether the user has a copy of
+  # this library. If not then add our copy's directory to $LD_LIBRARY_PATH.
 done
 
-for name in "${extracted_appimages[@]}"; do
-  symlink="$(which "${name}")"
-  apprun="$(dirname "${symlink}")/$(readlink "${symlink}")"
-  if [[ ! -L "${symlink}" || ! -f "${apprun}" ]]; then
-    echo "$0: Warning: Unable to find AppImage for '${name}'. Will not bundle." >&2
-    continue
+# pipewire is still packaged by linuxdeploy, 
+# so we need to remove it as it is in the fallback mechanism
+if [[ -f "${appdir}/lib/libpipewire-0.3.so.0" ]]; then
+  rm -f "${appdir}/lib/libpipewire-0.3.so.0"
+fi
+
+# APPIMAGEUPDATETOOL
+# Bundled uncompressed, to avoid creating a double layer of compression
+# (AppImage inside AppImage).
+if [[ "${UPDATE_INFORMATION}" ]]; then
+  if [[ ! -d $BUILD_TOOLS/appimageupdatetool ]]; then
+    mkdir $BUILD_TOOLS/appimageupdatetool
+    cd $BUILD_TOOLS/appimageupdatetool
+    download_appimage_release AppImageCommunity/AppImageUpdate appimageupdatetool continuous
+    cd $ORIGIN_DIR
   fi
-  extracted_appdir_path="$(dirname "${apprun}")"
-  extracted_appdir_name="$(basename "${extracted_appdir_path}")"
-  cp -r "${extracted_appdir_path}" "${appdir}/"
-  cat >"${appdir}/bin/${name}" <<EOF
+
+  export PATH="$BUILD_TOOLS/appimageupdatetool:$PATH"
+  appimageupdatetool --version
+
+  # Extract appimageupdatetool
+  appimageupdatetool --appimage-extract >/dev/null # dest folder "squashfs-root"
+
+  # Move into AppDir
+  mv squashfs-root ${appdir}/appimageupdatetool.AppDir
+
+  # Create alias in `${appdir}/bin`
+  # Use script instead of symlink, because appimageupdatetool.AppDir/AppRun fails
+  # when run from a symlink.
+  cat >"${appdir}/bin/appimageupdatetool" <<EOF
 #!/bin/sh
 unset APPDIR APPIMAGE # clear outer values before running inner AppImage
 HERE="\$(dirname "\$(readlink -f "\$0")")"
-exec "\${HERE}/../${extracted_appdir_name}/AppRun" "\$@"
+exec "\${HERE}/../appimageupdatetool.AppDir/AppRun" "\$@"
 EOF
-  chmod +x "${appdir}/bin/${name}"
-done
+  chmod +x "${appdir}/bin/appimageupdatetool"
+fi
 
 # METHOD OF LAST RESORT
 # Special treatment for some dependencies when all other methods fail
 
 # Bundle libnss3 and friends as fallback libraries. Needed on Chromebook.
 # See discussion at https://github.com/probonopd/linuxdeployqt/issues/35
-libnss3_files=(
-  # https://packages.ubuntu.com/xenial/amd64/libnss3/filelist
-  libnss3.so
-  libnssutil3.so
-  libsmime3.so
-  libssl3.so
-  nss/libfreebl3.chk
-  nss/libfreebl3.so
-  nss/libfreeblpriv3.chk
-  nss/libfreeblpriv3.so
-  nss/libnssckbi.so
-  nss/libnssdbm3.chk
-  nss/libnssdbm3.so
-  nss/libsoftokn3.chk
-  nss/libsoftokn3.so
-)
-
 libnss3_system_path="$(dirname "$(find_library libnss3.so)")"
 libnss3_appdir_path="${appdir}/fallback/libnss3.so" # directory named like library
 
-mkdir -p "${libnss3_appdir_path}/nss"
+mkdir -p "${libnss3_appdir_path}"
+
+if [ -d "${libnss3_system_path}/nss" ]; then
+  mkdir -p "${libnss3_appdir_path}/nss"
+
+  libnss3_files=(
+    # https://packages.ubuntu.com/jammy/amd64/libnss3/filelist
+    libnss3.so
+    libnssutil3.so
+    libsmime3.so
+    libssl3.so
+    nss/libfreebl3.chk
+    nss/libfreebl3.so
+    nss/libfreeblpriv3.chk
+    nss/libfreeblpriv3.so
+    nss/libnssckbi.so
+    nss/libnssdbm3.chk
+    nss/libnssdbm3.so
+    nss/libsoftokn3.chk
+    nss/libsoftokn3.so
+  )
+else
+  libnss3_files=(
+    # https://packages.ubuntu.com/noble/amd64/libnss3/filelist
+    libfreebl3.chk
+    libfreebl3.so
+    libfreeblpriv3.chk
+    libfreeblpriv3.so
+    libnss3.so
+    libnssckbi.so
+    libnssdbm3.chk
+    libnssdbm3.so
+    libnssutil3.so
+    libsmime3.so
+    libsoftokn3.chk
+    libsoftokn3.so
+    libssl3.so
+  )
+fi
 
 for file in "${libnss3_files[@]}"; do
   cp -L "${libnss3_system_path}/${file}" "${libnss3_appdir_path}/${file}"
@@ -330,15 +344,5 @@ fi
 
 # create AppImage
 appimagetool "${appimagetool_args[@]}" "${appdir}" "${appimage}"
-
-# We are running as root in the Docker image so all created files belong to
-# root. Allow non-root users outside the Docker image to access these files.
-chmod a+rwx "${created_files[@]}"
-parent_dir="${PWD}"
-while [[ "$(dirname "${parent_dir}")" != "${parent_dir}" ]]; do
-  [[ "$parent_dir" == "/" ]] && break
-  chmod a+rwx "$parent_dir"
-  parent_dir="$(dirname "$parent_dir")"
-done
 
 echo "Making AppImage finished"
